Refactor pluto run to startup file
