+++ src/main/globals.ts
@@ -1,0 +1,6 @@
+export class Globals {
+  public static JULIA: string;
+  public static JULIA_PROJECT: string;
+  public static PLUTO_LOCATION: string;
+  public static PLUTO_SECRET: string;
+}
